feature-auth/pyproject.toml
[project]
name = "ff-codex-api"
version = "0.1.0"
description = "Fantasy Football Codex API"
authors = [
    {name = "Jim fantasyfootballcodex@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.17.1"

